checkpoint: current sources in new repo
This includes local changes to:
1. Add prototypes/fix warnings.
2. Use dedicated lock files (avoids lock bombs)
3. Use select instead of alarms
4. A lot of old, dead code can be removed from party, so remove it.
5. Update utmp logic to work on FreeBSD

Party now uses a separate lock file in a protected
directory that normal users can't get access to.  This
prevents the sort of abuse where a user might open the
channel log and then take a lock on it, thus preventing
users from appending to the party log.

Users may still do so but it's irrelevant, since the
channel log itself is no longer part of the  locking
protocol on writes.

Signed-off-by: Dan Cross <cross@m-net.arbornet.org>